projects/195-time-chatbot/requirements.txt
openai
python-dotenv
flask
flask-cors
langchain
langchain-openai
langchain-core